# public/openapi.yaml
openapi: 3.1.0
info:
  title: tevaxia.lu API
  version: 1.1.1
  description: >-
    API de calculs indicatifs. Les hypothèses internes ne constituent ni une expertise certifiée ni une
    décision de crédit. Authentification selon chaque endpoint, par X-API-Key ou Authorization: Bearer. Les
    endpoints authentifiés appliquent des quotas et peuvent retourner 429 avec Retry-After.
    Pour les clés API, les limites sont partagées par clé sur une même instance serveur : Free 10/minute
    et 200/24 heures, Pro 60/minute et 5000/24 heures, Enterprise 600/minute et 100000/24 heures.
    Les fenêtres démarrent au premier appel ; ces compteurs en mémoire ne constituent pas un quota
    global distribué. La clé sandbox partage le quota Free entre ses utilisateurs. Une indisponibilité
    de la vérification des clés peut retourner 503.
  contact:
    name: Support tevaxia.lu
    email: contact@tevaxia.lu
    url: https://www.tevaxia.lu
  license:
    name: Proprietary
servers:
  - url: https://tevaxia.lu
    description: Production
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
paths:
...
  /api/v1/ai/analyze:
    post:
      tags:
        - AI
      summary: Analyse IA d'un résultat (commentaire expert)
      description: |
        Génère un commentaire professionnel par LLM (Cerebras Llama 3.3
        par défaut, ou BYOK OpenAI/Anthropic) à partir d'un contexte
        structuré (inputs + résultats d'un calcul) et d'un prompt métier.
        Rate-limit : 5/jour gratuit pour JWT, tier clé API pour X-API-Key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - context
                - prompt
              properties:
                context:
                  type: string
                  description: Inputs + résultats formatés (plain text multiligne)
                prompt:
                  type: string
                  description: Consigne métier pour le LLM
      responses:
        '200':
          description: Analyse IA générée
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                  model:
                    type: string
                  provider:
                    type: string
                    enum:
                      - cerebras
                      - groq
                      - openai
                      - anthropic
                  remaining:
                    type: integer
                    description: '-1 si illimité'
        '401':
          description: Auth requise
        '429':
          description: Quota quotidien atteint
        '502':
          description: Erreur fournisseur IA
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
...